+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add relation list fields @robgietema
 - Add @export view to export current content object data @robgietema
 - Add exclude from nav to type information @robgietema
+- Move exclude form nav by type to controlpanel @robgietema
 
 ### Bugfix